\name{bcnt.emapw}
\alias{bcnt.emapw}
\docType{data}
\title{Benthic count data for the western United States}
\description{
  Benthic count data for streams in the western United States. 
}
\format{
  Three fields are included.
  \describe{
  \item{\code{ID.NEW}}{Unique sample code}
  \item{\code{Taxaname}}{Name of the observed taxon}
  \item{\code{Abundance}}{Abundance of the observed taxon}
}
}
\usage{data(bcnt.emapw)}
\source{
  U.S. Environmental Protection Agency Environmental Monitoring and
  Assessment Program
}
\examples{
data(bcnt.emapw)
}
\keyword{datasets}
