\name{bcnt.OR}
\alias{bcnt.OR}
\docType{data}
\title{Benthic count data for western Oregon}
\description{
  Benthic count data for streams in western Oregon. 
}
\format{
  Three fields are included.
  \describe{
  \item{\code{SVN}}{Unique sample code}
  \item{\code{Taxon}}{Name of the observed taxon}
  \item{\code{CountValue}}{Abundance of the observed taxon}
}
}
\usage{data(bcnt.OR)}
\source{
  Oregon Department of Environmental Quality
}
\examples{
data(bcnt.OR)
}
\keyword{datasets}
