\name{bcnt.tax.OR}
\alias{bcnt.tax.OR}
\docType{data}
\title{Benthic count with taxonomic hierarchy}
\description{
  Benthic count data from western Oregon with full taxonomic hierarchy
  assigned by \code{\link{get.taxonomic}}.
}
\usage{data(bcnt.tax.OR)}
\source{
  Oregon Department of Environmental Quality
}
\examples{
data(bcnt.tax.OR)
}
\keyword{datasets}
