\name{envdata.emapw}
\alias{envdata.emapw}
\docType{data}
\title{Environmental data from the western United States}
\description{
  Environmental data for stream sampled in the western United States.
}
\usage{data(envdata.OR)}
\format{
  A data frame with 1674 observations on the following 8 variables.
  \describe{
    \item{\code{ID.NEW}}{Unique sample code}
    \item{\code{ntl}}{Total nitrogen concentration}
    \item{\code{no3}}{Nitrate concentration}
    \item{\code{ptl}}{Total phosphorus concentration}
    \item{\code{sed}}{Percent sands and fines}
    \item{\code{tempres}}{Corrected temperature}
    \item{\code{STRMTEMP}}{Grab temperature}
    \item{\code{jday}}{Julian sampling day}
  }
}
\source{
  U.S. Environmental Protection Agency Environmental Monitoring and
  Assessment Program.
}
\examples{
data(envdata.emapw)
}
\keyword{datasets}
