\name{flist.match}
\alias{flist.match}
\title{Selects data files matching search string}
\description{
  Select data files matching search string
}
\usage{
flist.match(pattern = NULL)
}
\arguments{
  \item{pattern}{Character string to be matched.}
}
\value{
  List of matched data files.
}
\author{Lester L. Yuan}

\keyword{models}

