\name{get.taxon.names}
\alias{get.taxon.names}
\title{Get taxon names from benthic count file}
\description{
Get taxon names from benthic count file
}
\usage{
get.taxon.names(bcnt)
}
\arguments{
  \item{bcnt}{Benthic count file with taxon names in second column}
}
\details{
  Checks to see if second column is factor or character.
}
\value{
  Returns character vector of taxon names.
}
\author{Lester L. Yuan}
\keyword{models}
