\name{incorp.correct}
\alias{incorp.correct}
\title{Incorporate taxon name revisions}
\description{
Incorporates corrected taxon names into parse.list
}
\usage{
incorp.correct(tname.new, parse.list)
}
\arguments{
  \item{tname.new}{Character vector of corrected taxon names}
  \item{parse.list}{List of parsed taxon names}
}
\value{
Revised parse.list.
}
\author{Lester L. Yuan}
\keyword{models}
