\name{infergui}
\alias{infergui}

\title{Graphical User Interface for biological inferences}

\description{
  Provides a graphical user interface that allows users to compute
  biological inferences using existing taxon-environment relationships.
}

\usage{
infergui()
}

\value{
  None.  Used for its side effects.
}

\author{
Lester L. Yuan
}
\keyword{models}
