\name{itis.ttable}
\alias{itis.ttable}
\docType{data}
\title{ITIS taxonomic hierarchy table}
\description{
  Full taxonomic hierarchy to genus for the Kingdom Animalia.
  Retrieved November 21, 2008 from the Integrated Taxonomic Information
  System on-line database, \url{http://www.itis.gov}.
}
\usage{data(itis.ttable)}
\source{
  Integrated Taxonomic Information System, U.S. Department of
  Agriculture.
}
\examples{
data(itis.ttable)
}
\keyword{datasets}
