\name{load.itis}
\alias{load.itis}
\title{Load ITIS taxon table}
\description{
  Load ITIS taxon table to the specified environment.
}
\usage{
load.itis(get.tax.env)
}
\arguments{
  \item{get.tax.env}{Environment variable.}
}
\details{
  Loads full ITIS taxonomic table into specified environment.
}
\value{
  None.  Used for side effects.
}
\author{Lester L. Yuan}
\keyword{models}

