\name{output.tax.table}
\alias{output.tax.table}
\title{Output final taxonomic table}
\description{
Output final taxonomic table as tab-delimited text
}
\usage{
output.tax.table(finaltab, tlevs)
}
\arguments{
  \item{finaltab}{Merged taxonomic table}
  \item{tlevs}{Character vector of taxonomic levels in ITIS}
}
\details{
  Outputs final table to file.
}
\value{
  None. Used for side effects.
}
\author{Lester L. Yuan}
\keyword{models}
