\name{sel.coeffile}
\alias{sel.coeffile}
\title{Select coefficient file}
\description{
Prompt user to select coefficient file from those available in data directory
}
\usage{
sel.coeffile(fstring = "coef")
}
\arguments{
  \item{fstring}{Character string for filtering files in data directory}
}
\details{
  Provides the list of data files available to the user.
}
\value{
  Selected file name.
}
\author{Lester L. Yuan}
\keyword{models}
