\name{ss.OR}
\alias{ss.OR}
\docType{data}
\title{site-OTU matrix for western Oregon}
\description{
  Matrix with each row representing a distinct sample and each colum
  representing a different OTU for western Oregon benthic count data.
}
\usage{data(ss.OR)}
\source{
  Oregon Department of Environmental Quality
}
\examples{
data(ss.OR)
}
\keyword{datasets}
