\name{trait.feeding}
\alias{trait.feeding}
\docType{data}
\title{Feeding traits for benthic invertebrates}
\description{
  Feeding classifications for different benthic invertebrates. 
}
\usage{data(trait.feeding)}
\format{
  A data frame with two columns.
  \describe{
    \item{\code{TAXON}}{Character vector with the names of each taxon.}
    \item{\code{TRAITVAL}}{Value of trait classification}
  }
}
\details{
  These trait classifications were summarized from the USGS data set
  prepared by Viera et al. \url{http://pubs.usgs.gov/ds/ds187/}.
}
\source{
  Yuan, L.L.
}
\examples{
data(trait.feeding)
}
\keyword{datasets}
